requirements.txt
Package            Version
------------------ ---------
certifi            2021.10.8
charset-normalizer 2.0.12
click              8.1.3
colorama           0.4.4
Flask              2.1.2
Flask-Login        0.6.1
greenlet           1.1.2
idna               3.3
importlib-metadata 4.11.3
itsdangerous       2.1.2
Jinja2             3.1.2
MarkupSafe         2.1.1
Pillow             9.1.0
pip                21.1.2
requests           2.27.1
setuptools         57.0.0
SQLAlchemy         1.4.36
urllib3            1.26.9
Werkzeug           2.1.2
wheel              0.36.2
zipp               3.8.0
